Add compaction validation trait for compaction schedulers
